snapshots: Minor refactoring
Move NVMe admin command handler to the nvmf subsystem.

In Replica, move the snapshot done callback to the function that uses
it, refactor use of nvme_status, and use ffihelper for conversion to a
C string.

Add function comments where appropriate.

In the integration tests, factor out read_some and write_some to
common and swtich current users to that.

Fixes: d909c1e ("replica: Implement create_snapshot")
