+++ mayastor/src/bdev/nexus/nexus_bdev.rs
@@ -33,7 +33,6 @@ use spdk_sys::{
     spdk_io_channel,
     spdk_io_device_register,
     spdk_io_device_unregister,
-    spdk_nvmf_request,
 };
 
 use crate::{
@@ -54,7 +53,6 @@ use crate::{
     ffihelper::errno_result_from_i32,
     nexus_uri::{bdev_destroy, NexusBdevError},
     rebuild::RebuildError,
-    replica::Replica,
 };
 
 /// Obtain the full error chain
@@ -372,60 +370,6 @@ impl Drop for Nexus {
             let _ = std::ffi::CString::from_raw(b.name);
             let _ = std::ffi::CString::from_raw(b.product_name);
         }
-    }
-}
-
-extern "C" fn nvmf_create_snapshot_hdlr(req: *mut spdk_nvmf_request) -> i32 {
-    debug!("nvmf_create_snapshot_hdlr {:?}", req);
-
-    let subsys = unsafe { spdk_sys::spdk_nvmf_request_get_subsystem(req) };
-    if subsys.is_null() {
-        debug!("subsystem is null");
-        return -1;
-    }
-
-    /* Only process this request if it has exactly one namespace */
-    if unsafe { spdk_sys::spdk_nvmf_subsystem_get_max_nsid(subsys) } != 1 {
-        debug!("multiple namespaces");
-        return -1;
-    }
-
-    /* Forward to first namespace if it supports NVME admin commands */
-    let mut bdev: *mut spdk_bdev = std::ptr::null_mut();
-    let mut desc: *mut spdk_bdev_desc = std::ptr::null_mut();
-    let mut ch: *mut spdk_io_channel = std::ptr::null_mut();
-    let rc = unsafe {
-        spdk_sys::spdk_nvmf_request_get_bdev(
-            1, req, &mut bdev, &mut desc, &mut ch,
-        )
-    };
-    if rc != 0 {
-        /* No bdev found for this namespace. Continue. */
-        debug!("no bdev found");
-        return -1;
-    }
-
-    let bd = Bdev::from(bdev);
-    if let Some(replica) = Replica::from_bdev(&bd) {
-        let cmd = unsafe { &*spdk_sys::spdk_nvmf_request_get_cmd(req) };
-        let snapshot_time = unsafe {
-            cmd.__bindgen_anon_1.cdw10 as u64
-                | (cmd.__bindgen_anon_2.cdw11 as u64) << 32
-        };
-        let snapshot_name = format!("{}-snap-{}", bd.name(), snapshot_time);
-        replica.create_snapshot(req, &snapshot_name);
-        1 // SPDK_NVMF_REQUEST_EXEC_STATUS_ASYNCHRONOUS
-    } else {
-        -1
-    }
-}
-
-pub fn setup_create_snapshot_hdlr() {
-    unsafe {
-        spdk_sys::spdk_nvmf_set_custom_admin_cmd_hdlr(
-            nvme_admin_opc::CREATE_SNAPSHOT,
-            Some(nvmf_create_snapshot_hdlr),
-        );
     }
 }
