+++ mayastor/tests/common/mod.rs
@@ -12,8 +12,10 @@ use mayastor::{
 };
 use spdk_sys::spdk_get_thread;
 
+pub mod bdev_io;
 pub mod error_bdev;
 pub mod ms_exec;
+
 /// call F cnt times, and sleep for a duration between each invocation
 pub fn retry<F, T, E>(mut cnt: u32, timeout: Duration, mut f: F) -> T
 where
