+++ mayastor/tests/io.rs
@@ -1,13 +1,8 @@
 use std::process::Command;
 
+use common::bdev_io;
 use mayastor::{
-    core::{
-        mayastor_env_stop,
-        Bdev,
-        MayastorCliArgs,
-        MayastorEnvironment,
-        Reactor,
-    },
+    core::{mayastor_env_stop, MayastorCliArgs, MayastorEnvironment, Reactor},
     nexus_uri::bdev_create,
 };
 
@@ -46,41 +41,7 @@ fn io_test() {
 // only execute one future per reactor loop.
 async fn start() {
     bdev_create(BDEVNAME).await.expect("failed to create bdev");
-    write_some().await;
-    read_some().await;
+    bdev_io::write_some(BDEVNAME).await.unwrap();
+    bdev_io::read_some(BDEVNAME).await.unwrap();
     mayastor_env_stop(0);
-}
-
-async fn write_some() {
-    let bdev = Bdev::lookup_by_name(BDEVNAME).expect("failed to lookup bdev");
-    let d = bdev
-        .open(true)
-        .expect("failed open bdev")
-        .into_handle()
-        .unwrap();
-    let mut buf = d.dma_malloc(512).expect("failed to allocate buffer");
-    buf.fill(0xff);
-
-    let s = buf.as_slice();
-    assert_eq!(s[0], 0xff);
-
-    d.write_at(0, &buf).await.unwrap();
-}
-
-async fn read_some() {
-    let bdev = Bdev::lookup_by_name(BDEVNAME).expect("failed to lookup bdev");
-    let d = bdev.open(false).unwrap().into_handle().unwrap();
-    let mut buf = d.dma_malloc(1024).expect("failed to allocate buffer");
-    let slice = buf.as_mut_slice();
-
-    assert_eq!(slice[0], 0);
-    slice[512] = 0xff;
-    assert_eq!(slice[512], 0xff);
-
-    d.read_at(0, &mut buf).await.unwrap();
-
-    let slice = buf.as_slice();
-
-    assert_eq!(slice[0], 0xff);
-    assert_eq!(slice[512], 0);
 }
